nivel de acceso verificado en base a los permisos

+++ server/routes/users.js
@@ -58,13 +58,13 @@ app.post('/users', (req, res) => {
                         } else {
                             res.status(400).json({
                                 err: {
-                                    message: "El nombre del uuario y contraseña son necesarios"
+                                    message: "El nombre del usuario y contraseña son necesarios"
                                 }
                             });
                         }
                     }else{
                         return res.status(403).json({
-                            message: 'No esta autorizado para crear usuarios'
+                            message: 'No está autorizado para crear usuarios'
                         });
                     }   
                 }
@@ -80,8 +80,11 @@ app.post('/users', (req, res) => {
                     contra: bcrypt.hashSync(body.contra, 10),
                     recuperacion: body.ci,
                     permisos: ['es_leer', 'es_escribir', 'es_borrar', 'es_modificar',
-                            'or_leer', 'or_escribir', 'or_borrar', 'or_modificar',
-                            'u_leer', 'u_escribir', 'u_borrar', 'u_modificar']
+                               'or_leer', 'or_escribir', 'or_borrar', 'or_modificar',
+                               'u_leer', 'u_escribir', 'u_borrar', 'u_modificar', 
+                               'ru_leer', 'ru_escribir', 'ru_borrar', 'ru_modificar',
+                               'p_leer', 'p_escribir', 'p_borrar', 'p_modificar',
+                               've_leer', 've_escribir', 've_borrar', 've_modificar']
                 });
                 user.save((err, userDB) => {
                     if (err) {
@@ -96,7 +99,7 @@ app.post('/users', (req, res) => {
             } else {
                 res.status(400).json({
                     err: {
-                        message: "El nombre del uuario y contraseña son necesarios"
+                        message: "El nombre del usuario y contraseña son necesarios"
                     }
                 });
             }
@@ -109,16 +112,25 @@ app.post('/users', (req, res) => {
 // ===============================================
 app.get('/users/:id', verifyToken, (req, res) => {
     let id = req.params.id;
-    Usuario.findById(id, (err, dbUser) => {
-        if (err) {
-            return res.status(500).json({
-                err
+    let user = decoded.user;
+    if(user.permisos.includes('u_leer')){
+        Usuario.findById(id, (err, dbUser) => {
+            if (err) {
+                return res.status(500).json({
+                    err
+                });
+            }
+            res.json({
+                user: dbUser
             });
-        }
-        res.json({
-            user: dbUser
         });
-    });
+    }else{
+        res.status(403).json({
+            err: {
+                message: 'No está autorizado para observar usuarios'
+            }
+        });
+    }
 });
 
 // ===============================================
@@ -129,23 +141,32 @@ app.get('/users/:id', verifyToken, (req, res) => {
 app.get('/users', verifyToken, (req, res) => {
     let from = Number(req.query.from) || 0;
     let limit = Number(req.query.to) || 15;
-    // TODO: define search params
-    Usuario.find({}, 'nombreUsuario permisos empresa')
-        .skip(from)
-        .limit(limit)
-        .exec((err, users) => {
-            if (err) {
-                return res.status(400).json({
-                    err
-                });
-            }
-            Usuario.countDocuments({}, (err, c) => {
-                res.json({
-                    users,
-                    count: c
+    let user = decoded.user;
+    if(user.permisos.includes('u_leer')){
+        // TODO: define search params
+        Usuario.find({}, 'nombreUsuario permisos empresa')
+            .skip(from)
+            .limit(limit)
+            .exec((err, users) => {
+                if (err) {
+                    return res.status(400).json({
+                        err
+                    });
+                }
+                Usuario.countDocuments({}, (err, c) => {
+                    res.json({
+                        users,
+                        count: c
+                    });
                 });
             });
+    }else{
+        res.status(403).json({
+            err: {
+                message: 'No está autorizado para observar usuarios'
+            }
         });
+    }
 });
 
 // ===============================================
@@ -154,21 +175,28 @@ app.get('/users', verifyToken, (req, res) => {
 app.put('/users/:id', verifyToken, (req, res) => {
     let body = _.pick(req.body, ['permisos', 'contra', 'empresa']);
     let id = req.params.id;
-
-    if (body.contra) {
-        body.contra = bcrypt.hashSync(body.contra, 10);
-    }
-
-    Usuario.findByIdAndUpdate(id, body, { new: true, runValidators: true }, (err, dbUsuario) => {
-        if (err) {
-            return res.status(500).json({
-                err
-            });
+    let user = decoded.user;
+    if(user.permisos.includes('u_modificar')){
+        if (body.contra) {
+            body.contra = bcrypt.hashSync(body.contra, 10);
         }
-        res.json({
-            user: dbUsuario
+        Usuario.findByIdAndUpdate(id, body, { new: true, runValidators: true }, (err, dbUsuario) => {
+            if (err) {
+                return res.status(500).json({
+                    err
+                });
+            }
+            res.json({
+                user: dbUsuario
+            });
         });
-    });
+    }else{
+        res.status(403).json({
+            err: {
+                message: 'No está autorizado para modificar usuarios'
+            }
+        });
+    }
 });
 
 // ===============================================
@@ -176,24 +204,32 @@ app.put('/users/:id', verifyToken, (req, res) => {
 // ===============================================
 app.delete('/users/:id', verifyToken, (req, res) => {
     let id = req.params.id;
-
-    Usuario.findByIdAndRemove(id, (err, deletedUser) => {
-        if (err) {
-            return res.status(500).json({
-                err
+    let user = decoded.user;
+    if(user.permisos.includes('u_borrar')){
+        Usuario.findByIdAndRemove(id, (err, deletedUser) => {
+            if (err) {
+                return res.status(500).json({
+                    err
+                });
+            }
+            if (!deletedUser) {
+                return res.status(400).json({
+                    err: {
+                        message: 'Usuario no encontrado'
+                    }
+                });
+            }
+            res.json({
+                message: `La cuenta de ${deletedUser.nombreUsuario} ha sido eliminada`
             });
-        }
-        if (!deletedUser) {
-            return res.status(400).json({
-                err: {
-                    message: 'Usuario no encontrado'
-                }
-            });
-        }
-        res.json({
-            message: `La cuenta de ${deletedUser.nombreUsuario} ha sido eliminada`
         });
-    });
+    }else{
+        res.status(403).json({
+            err: {
+                message: 'No está autorizado para borrar usuarios'
+            }
+        });
+    }
 });
